+++ resources/views/components/sidebar.blade.php
@@ -1,7 +1,7 @@
 <div data-sidebar class="fixed inset-0" style="z-index: 50; display: none;">
     <div data-sidebar-backdrop data-sidebar-close class="absolute inset-0 bg-black opacity-25 fade-in"></div>
 
-    <aside data-sidebar-content class="absolute top-0 right-0 bottom-0 bg-white w-1/2 shadow-lg p-16 slide-from-right">
+    <aside data-sidebar-content class="absolute top-0 right-0 bottom-0 bg-white shadow-lg p-16 slide-from-right overflow-auto">
         <div data-sidebar-close class="absolute top-0 right-0 m-4 cursor-pointer">
             <div class="rounded-full p-1 text-grey-500 bg-grey-50 hover:bg-grey-100 transition duration-150 ease-in-out">
                 <svg width="16" height="16"><use xlink:href="#x"></use></svg>