+++ src/Fragments/Database/ContextModel.php
@@ -4,25 +4,24 @@ declare(strict_types=1);
 namespace Thinktomorrow\Chief\Fragments\Database;
 
 use Illuminate\Database\Eloquent\Model;
-use Thinktomorrow\Chief\Fragments\FragmentsOwner;
 
 final class ContextModel extends Model
 {
     public $table = "context";
     public $guarded = [];
 
-    public static function ownedBy(FragmentsOwner $owner): ?ContextModel
+    public static function ownedBy(Model $owner): ?ContextModel
     {
-        return static::where('owner_type', $owner->modelReference()->className())
-                     ->where('owner_id', $owner->modelReference()->id())
+        return static::where('owner_type', $owner->getMorphClass())
+                     ->where('owner_id', $owner->id)
                      ->first();
     }
 
-    public static function createForOwner(FragmentsOwner $owner): ContextModel
+    public static function createForOwner(Model $owner): ContextModel
     {
         return static::create([
-            'owner_type' => $owner->modelReference()->className(),
-            'owner_id' => $owner->modelReference()->id(),
+            'owner_type' => $owner->getMorphClass(),
+            'owner_id' => $owner->id,
         ]);
     }